separated message components and removed createmessagemodal
