+++ react-app/src/components/CreateMessage/index.js
@@ -1,0 +1,3 @@
+import CreateMessage from "./CreateMessage";
+
+export default CreateMessage;
