+++ react-app/src/components/MessageCard/index.js
@@ -1,0 +1,3 @@
+import MessageCard from "./MessageCard";
+
+export default MessageCard;
