--- react-app/src/components/Messages/CreateMessage.js
@@ -1,14 +1,0 @@
-import React, { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-
-export default function CreateMessage() {
-    const dispatch = useDispatch();
-
-    return (
-        <div>
-            <select>
-                <option>---Select a Friend---</option>
-            </select>
-        </div>
-    );
-}
