--- react-app/src/components/Messages/CreateMessageModal.module.css
@@ -1,8 +1,0 @@
-.wrapper {
-    display: flex;
-}
-
-.button {
-    max-height: 35px;
-    max-width: 35px;
-}
